src/main/java/User/User.java
package User;

import java.io.Serializable;
import java.util.Objects;

public class User implements Serializable {

    int id;
    String name;
    String email;

    public User() {
    }

    public User(int id, String name, String email) {
        this.id = id;
        this.name = name;
        this.email = email;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) return true;
        if (obj == null || getClass() != obj.getClass()) return false;
        User user = (User) obj;
        return id == user.id && Objects.equals(name, user.name) && Objects.equals(email, user.email);
    }

    @Override
    public String toString() {
        return "User: " + "id= " + id + ", name= " + name + ", email= " + email + '|';
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name, email);
    }


    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}
